Split login from server
